=== != what you think it is unless it’s what you think it is on both sides.

+++ inc/class-sitemaps.php
@@ -477,7 +477,7 @@ class WPSEO_Sitemaps {
 				$p->post_status = 'publish';
 				$p->filter      = 'sample';
 
-				if ( wpseo_get_value( 'meta-robots-noindex', $p->ID ) === 1 && wpseo_get_value( 'sitemap-include', $p->ID ) != 'always' ) {
+				if ( (int) wpseo_get_value( 'meta-robots-noindex', $p->ID ) === 1 && wpseo_get_value( 'sitemap-include', $p->ID ) != 'always' ) {
 					continue;
 				}
 				if ( wpseo_get_value( 'sitemap-include', $p->ID ) == 'never' ) {
